Add stream-consuming middleware.
According to the uWSGI docs:

> If an HTTP request has a body (like a POST request generated by a
  form), you have to read (consume) it in your application. If you do
  not do this, the communication socket with your webserver may be
  clobbered. [1]

StreamConsumingMiddleware ensures that the body is consumed. Without
it, in cases where a large request body is streamed but (for example)
authentication fails, communication between nginx and uWSGI fails,
resulting in a 502 Bad Gateway.

Unfortunately, consuming the incoming stream properly in this way
confuses the Werkzeug test client in cases where no content is
returned (as is often the case with updates). So in these cases we
have to pass `buffered=True` to make the test client work properly.
that means.

[1] http://uwsgi-docs.readthedocs.io/en/latest/ThingsToKnow.html