Use server-specific URLs in history
Changes and other history-related entities are specific to a dataset
and shouldn't receive permanent URLs unless the server is
canonical. These changes also unify how the context is handled for the
dataset and the history. As a side effect they break the RDF
interpretation of the dataset until a data patch adding the PROV terms
to the context is applied (see the periodo-validation repo).