In ORCID authentication, determine client host via a query parameter
The client's host must be known in order to send it a message via
postMessage. Prior to this commit, it was hardcoded in the global
variable CLIENT_URL.

This change makes the server agnostic about what the host of a client
is, except as a default value. (Which could be taken out).